requirements.txt
altair==3.2.0
argh==0.26.2
astor==0.8.0
attrs==19.3.0
base58==1.0.3
blinker==1.4
boto3==1.10.8
botocore==1.13.8
certifi==2019.9.11
chardet==3.0.4
Click==7.0
decorator==4.4.1
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
flake8==3.7.9
future==0.18.2
graphviz==0.13
idna==2.8
importlib-metadata==0.23
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==3.1.1
lxml==4.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.17.3
pandas==0.25.3
pathtools==0.1.2
Pillow==6.2.1
protobuf==3.10.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyrsistent==0.15.5
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
streamlit==0.49.0
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
tzlocal==2.0.0
urllib3==1.25.6
validators==0.14.0
watchdog==0.9.0
zipp==0.6.0
